Refactor startup subsystems into modules
Split AirGap startup logic into dedicated modules and clean up related code:

- Added lib/auto_start.py, lib/crash_recovery.py, lib/update_check.py and lib/startup_common.py to encapsulate auto-start, crash-recovery, update-check and shared startup-wait logic (scheduling, event handlers, readiness checks).
- Updated AirGap.py to use schedule_* and cleanup functions from the new modules, removed the large inline startup/crash/update implementations and simplified shutdown cleanup.
- Added CUSTOM_EVENT_* constants to config.py for the new custom events.
- Refactored autosave manager: added activate_if_enabled helper, extracted export/record logic for clearer error handling and reduced duplication.
- Simplified settings command UI code by extracting autosave inputs and handler registration, and consolidated folder-browse handling.
- start_session now uses autosave_manager.activate_if_enabled to start autosave when a session begins.

Overall this modularization reduces duplication, improves separation of concerns, and makes startup-related behavior easier to maintain and test.
